- hotels and room controllers with inventory management added

+++ src/main/java/com/example/airbnb/entity/Hotel.java
@@ -1,5 +1,6 @@
 package com.example.airbnb.entity;
 
+import com.example.airbnb.dto.HotelDTO;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -18,7 +19,7 @@ import java.util.List;
 public class Hotel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false)
     private String name;
